Break apart utils into eth::utils and extensions::{timeout, removed}
