+++ src/eth/mod.rs
@@ -1,5 +1,6 @@
 pub mod contracts;
 pub mod transaction;
+pub mod utils;
 
 use super::errors;
 use super::relay;
