+++ src/eth/utils.rs
@@ -1,0 +1,8 @@
+// From ethereum_types but not reexported by web3
+pub fn clean_0x(s: &str) -> &str {
+    if s.starts_with("0x") {
+        &s[2..]
+    } else {
+        s
+    }
+}
