+++ src/extensions/timeout.rs
@@ -1,23 +1,10 @@
 use std::ops::Add;
-use std::rc::Rc;
 use std::time::{Duration, Instant};
 use tokio_core::reactor;
 use web3::api::{SubscriptionResult, SubscriptionStream};
 use web3::futures::prelude::*;
 use web3::futures::try_ready;
-use web3::types::H256;
 use web3::{DuplexTransport, ErrorKind};
-
-use relay::{Network, TransferApprovalState};
-
-// From ethereum_types but not reexported by web3
-pub fn clean_0x(s: &str) -> &str {
-    if s.starts_with("0x") {
-        &s[2..]
-    } else {
-        s
-    }
-}
 
 /// Enum for the two stages of subscribing to a timeout stream
 /// Subscribing is holds future that returns a TimeoutStream
@@ -153,63 +140,4 @@ where
         let handle = handle.clone();
         TimeoutStream::new(SubscriptionState::Subscribing(Box::new(self)), timeout, &handle)
     }
-}
-
-pub struct CheckLogRemoved<T, I, E>
-where
-    T: DuplexTransport + 'static,
-{
-    target: Rc<Network<T>>,
-    tx_hash: H256,
-    future: Box<Future<Item = I, Error = E>>,
-}
-
-impl<T, I, E> CheckLogRemoved<T, I, E>
-where
-    T: DuplexTransport + 'static,
-{
-    pub fn new(target: Rc<Network<T>>, tx_hash: H256, future: Box<Future<Item = I, Error = E>>) -> Self {
-        CheckLogRemoved {
-            target,
-            tx_hash,
-            future,
-        }
-    }
-}
-
-impl<T, I, E> Future for CheckLogRemoved<T, I, E>
-where
-    T: DuplexTransport + 'static,
-{
-    type Item = Option<I>;
-    type Error = E;
-
-    fn poll(&mut self) -> Poll<Self::Item, Self::Error> {
-        match self.future.poll() {
-            Ok(Async::Ready(result)) => Ok(Async::Ready(Some(result))),
-            Ok(Async::NotReady) => {
-                // Check removed status
-                match self.target.pending.read().unwrap().peek(&self.tx_hash) {
-                    Some(TransferApprovalState::Removed) => Ok(Async::Ready(None)),
-                    _ => Ok(Async::NotReady),
-                }
-            }
-            Err(e) => Err(e),
-        }
-    }
-}
-
-/// Trait to add to any Stream for creating a CheckLogRemoved via check_log_removed()
-pub trait CheckRemoved<T, I, E>
-where
-    T: DuplexTransport + 'static,
-{
-    ///Returns a TimeoutStream that wraps the existing stream
-    ///
-    /// # Arguments
-    ///
-    /// * `self` - Existing Stream that this is added to. Consumes self.
-    /// * `target` - Target network to check against
-    /// * `tx_hash` - Tx hash to check for removal
-    fn check_log_removed(self, target: &Rc<Network<T>>, tx_hash: H256) -> CheckLogRemoved<T, I, E>;
 }
