+++ src/server/endpoint.rs
@@ -18,11 +18,11 @@ use actix_web::http::{Method, StatusCode};
 use actix_web::{middleware, server, HttpResponse, Path};
 
 use super::errors::EndpointError;
+use super::eth::contracts::TRANSFER_EVENT_SIGNATURE;
+use super::eth::utils;
 use super::relay::{Network, NetworkType};
-use super::transactions::contracts::TRANSFER_EVENT_SIGNATURE;
 use super::transfers::past::ValidateAndApproveTransfer;
 use super::transfers::transfer::Transfer;
-use super::utils;
 
 pub const HOME: &str = "HOME";
 pub const SIDE: &str = "SIDE";
