+++ src/server/mod.rs
@@ -1,7 +1,6 @@
 pub mod endpoint;
 
 use super::errors;
+use super::eth;
 use super::relay;
-use super::transactions;
 use super::transfers;
-use super::utils;
