+++ src/transfers/mod.rs
@@ -3,7 +3,7 @@ pub mod past;
 pub mod transfer;
 pub mod withdrawal;
 
+use super::eth;
+use super::extensions;
 use super::relay;
-use super::transactions;
 use super::transfers;
-use super::utils;
